#78 [feat] stadium 어그리거트의 테스트 코드를 리팩토링 한 것에 맞게 수정한다.
